gitignore: add solution, project files to repo
I decided, after watching the second video in the tutorial ("Initial
Folder Setup", that I wanted to keep track of the solution and project
files.  There are a number of things in Visual Studio that are part of
my program organization, such as the virtual folders and external
dependencies and what projects are involved, that are stored in these
folders.