config adjustment didn't go through
I must have not checked the SandboxGame .vcxproj for commit last time.

Also, the PATH environment variable that I specified for the location of
the dlls is in the .user file, but it is generally recommended that the
.user file not be committed to a repo because it usually contains file
paths that are unique to your machine.

My guess is that the "additional include directories" were intended to
be relative paths within the repo.  Mine are absolute paths.  Oh well.