refactored tests and vector2D definitions
My tests were using, up to this point, the EXPECT_EQ(...) macro with
floating values.  I should have been using EXPECT_FLOAT_EQ(...), which
subtracts the two values, takes the absolute value, and checks if it is
below some threshold.  This takes into account the slightly imprecise
nature of floats.

I also removed the definitions of the vector2D operators from the header
file, re-specified the remaining declarations as "inline", and put the
operator function definitions into a ".inl" file that I included from
the header.  It's a bit crude, but it makes the code read more cleanly.