forced a hiccup in computations per frame
This hiccup is a demonstration of why it is not a good thing to base
rendering and game timing on how long it takes to process a frame.  If
there IS a computational hiccup somewhere, we need to keep some other
kind of timing mechanism going so that, when the heavy computation is
finished and other things are rendered, they draw the things in the
positions that they would have been in if the frame had not stuttered.