fixed cross-project linker errors
I instructed Visual Studio to export the declarations of the clock
class.  In doing this, the declared class wound up in a .lib file.  I
also changed the output directory of the Engine project's .dll file to
be it's Debug folder (technically I set it to the macro $(Configuration)
).  Then I changed the "Engine Tester" class to use these .lib and .dll
files.

It builds and links correctly now.