changed window update to use the clock
Previously, the ship position over time depended on the time to render
the frame.  The ship position over time is now dependent on the clock
class' calculated delta time.

Even with the lag spike that simulates a long calculation time between
frames, the ship's new position at drawing time is dependent only on the
clock and will appear where we expect it as if the lag spike had not
occurred.