some cleanup
I was getting compiler warnings about a performance issue in the clock
in converting the int return value of the query functions to a bool.  I
changed the code to be a true boolean comparison.

I also got a compiler warning that including .h files from the Qt
include\Qt\ directory was depreciated.  I changed to using their
equivalent non-header versions, and the warning went away.