changed matrix2D operator* to non-member function
I justified the operator* function being a member function so that I
would enforce the right-hand-sideness of the vector, but I later
realized that this was a bad plan.

The right-hand-sideness is still enforced by making the function a
non-member, and implicit construction of the member arguments is avoid
because both matrix and vector constructors are explicit.