renamed entries in matrix2D to be generic
Previously, I tried to have the names of each entry in the matrix be
intuitive, like x0, x1, x_translate, y0, y1, and y_translate, but the
third row was just sitting there and didn't have any intuitive name to
go with it.

Additionally, the intuitive names made matrix multiplication a really
big pain in the neck.

So I renamed them to row0_col0, row0_col1, ... row_2_col2.  I had to
remake my moc file because I did a text search and change in the entire
solution, and there were some hex values in the moc file that began with
x0, which the search picked up.  The moc file isn't part of this commit
though because I remade it so it appeared unchanged.

But now my rotations are all off.

I think that it is because the size of the vector changed.