integrated transformation matrix into render step
Previously, the transformation of the ship's vertices was performed by
applying the rotation matrix and then adding the ship position vector.

Now I create a rotation and translation matrix, multiply the translation
by the rotation (NOT the other way around, which will effectively
translate the ship, then rotate it around the origin) to create a
transformation matrix, then multiply the matrix by each vertex.

Woot!  It works too!