added matrix dimension for non-origin rotation
The names of the row_col variables were starting to get unweildy, so I
renamed them to a(row_num)(col_num).

Ex: Row 0, col 2 is a02.

This is a little terminology I picked up when looking online at matrix
multiplication.  It's simple enough.

Now I need to add the ability to rotate a non-origin axis and rewrite
the vector2D with the new dimension to enable non-origin rotation and
all the matrix and vector tests.  Ugh!