changed vector's last entry to "w"
I had been calling the vector2D's last value "t" because it enabled
translation, but I had forgotten that there is graphical programming
tradition that this value is called "w".  So I changed it to "w".

This "w" value is also supposed to not be restricted to 1.0f and 0.0f
only.  It has some usefulness in matrices in scaling the vector,
although it has little use to the vector itself.  The vector treats it
as any other entry, adding, subtracting, and scaling it.  It is ok to do
this because point ("w" = 1) and direction vectors ("w" = 0) may be
added, and direction and direction vectors may be added, but point and
point vectors will always remain 1 or 0 throughout vector manipulation
other than scaling or funky transforms.

Also, I changed the default "w" value from 1 to 0.