changed vertex declarations to enable translate
Before the last commit, the "w" (that is, "enable translate") value
defaulted to 1.  When I changed the way that "w" behaved so that it was
just like any other vector component, I changed it's default value to
("disable translate") as well, so I had to change the ship's vertex
declarations to explicitly enable translation as well.