cleaned up profiler and tests a bit
I changed the three discrete string literals for "Category1",
"Category2", and "Category3" into an array of string literals, and I
adjusted the rest of the test to match.

I also changed up some variable names to better convey intent.

After debugging it, the profiler as is works as expected, and its tests
pass again.