added profile test for an incomplete frame
I added a test that checked that, if a new frame was added, then that
data, even if some or all of it was 0, will be written to the file.

I also refactored the tests a little bit by moving the "fine and dandy"
profile writing to a single function and having both my tests call it.