made the profiler handle circular frame sampling
There are a limited number of samples that can be recorded per category,
but those will only last a few minutes at most if our frame count is
fixed at 30 or 60 fps.  We don't need every single piece of data though,
just the trend, so I rigged it up (and had to alter the test's "check
samples" function) to start overwriting the oldest values once it got to
the end of the category data structures.